requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
sqlalchemy>=2.0.0
python-multipart>=0.0.6
httpx>=0.25.0
aiofiles>=23.0.0
jinja2>=3.1.0

# Optional geospatial helpers
geopy>=2.4.0
haversine>=2.8.0
networkx>=3.2.0

# Load .env at startup
python-dotenv>=1.0.0
